Refactor. FavoriteResponse 패키지 이동
