+++ BE/src/main/java/team07/airbnb/data/user/dto/response/FavoritesResponse.java
@@ -1,4 +1,4 @@
-package team07.airbnb.data.user.dto.request;
+package team07.airbnb.data.user.dto.response;
 
 import team07.airbnb.data.product.dto.response.ProductListResponse;
